enhancement : Comment 좋아요 개수 구하는 쿼리 개선 및 Dto 수정

+++ dplanner/src/main/java/com/dp/dplanner/repository/CommentMemberLikeRepository.java
@@ -8,5 +8,5 @@ import java.util.Optional;
 public interface CommentMemberLikeRepository extends JpaRepository<CommentMemberLike,Long> {
     Optional<CommentMemberLike> findCommentMemberLikeByClubMemberIdAndCommentId(Long clubMemberId, Long commentId);
 
-    int countDistinctByCommentId(Long CommentId);
+    Long countDistinctByCommentId(Long CommentId);
 }
